requirements.txt
PyQt5==5.9.2
PyVISA==1.8
sip==4.19.6
